get -> request, fixed crawlers pt1, better logging

+++ fetcher/src/run-crawlers.js
@@ -1,49 +1,81 @@
-const CHILD_MAX_LIFETIME = 30 * 1000 * 60;
-
 const fs = require("fs");
 const { fork } = require("child_process");
 const async = require("async");
-const flatten = require("lodash.flatten");
 const path = require("path");
+const j = require("joi");
 
 const logger = require("./logger");
 const { parseEnvArray } = require("./utils");
 
 const crawlersDir = `${__dirname}/crawlers`;
+const MAX_TIME_BETWEEN_MSG = 10 * 60 * 1000;
+
+const DATE_SCHEMA = j.string().regex(/[0-9]{4}[-]{1}[0-9]{2}[-]{1}[0-9]{2}/g);
+
+const ENTITY_SCHEMA = j.object().keys({
+  title: j.string().required(),
+  sourceName: j.string().required(),
+  url: j.string().required(),
+  date: DATE_SCHEMA.required(),
+  updateDate: DATE_SCHEMA,
+  ocr: j.boolean()
+});
+
+const ENTITIES_SCHEMA = j.array().items(ENTITY_SCHEMA);
 
 // TODO: Test error handling
 module.exports = async ({ onEach }) => {
   const crawlersList = process.env.DEV_CRAWLERS ? parseEnvArray(process.env.DEV_CRAWLERS) : fs.readdirSync(crawlersDir);
 
-  return new Promise((resolve, reject) => {
-    async.mapSeries(
-      crawlersList,
-      (crawlerFileName, callback) => {
-        logger.info(`running ${crawlersDir}/${crawlerFileName}`);
-        const child = fork("./src/spawn-crawler", [path.join(crawlersDir, "/", crawlerFileName)]);
+  let timer;
 
-        child.on("error", err => {
-          callback(err);
-        });
+  let lastUpdateTime;
 
-        child.on("message", data => {
-          onEach(data);
-          callback(null, data);
-        });
-
-        // Kill if process takes longer than CHILD_MAX_LIFETIME
-        setTimeout(() => {
-          child.kill("SIGINT");
-        }, CHILD_MAX_LIFETIME);
-      },
-      (err, results) => {
-        if (err) {
-          logger.error(`Error in runCrawlers ${err}`);
-          return reject(err);
-        }
-
-        resolve(flatten(results));
+  const setNewTimer = child => {
+    if (timer) {
+      clearInterval(timer);
+    }
+    lastUpdateTime = Date.now();
+    setInterval(() => {
+      const now = Date.now();
+      if (now - lastUpdateTime >= MAX_TIME_BETWEEN_MSG) {
+        logger.error(`Process killed due to long time no response`, child);
+        child.kill("SIGINT");
       }
-    );
+    }, 60 * 1000);
+  };
+
+  async.eachLimit(crawlersList, 1, (crawlerFileName, callback) => {
+    const crawlerPath = path.join(crawlersDir, "/", crawlerFileName);
+    logger.info(`Running: ${crawlersDir}/${crawlerFileName}`);
+    const child = fork("./src/spawn-crawler", [crawlerPath]);
+
+    const startDate = Date.now();
+
+    setNewTimer(child);
+
+    child.on("error", err => {
+      setNewTimer(child);
+      logger.error(`Error in ${crawlerPath} - ${err}`);
+
+      callback(null);
+    });
+
+    child.on("message", data => {
+      setNewTimer(child);
+      j.validate(data, ENTITIES_SCHEMA, err => {
+        if (err) {
+          logger.warn(`The crawled entity is not valid with the schema. Error: ${err}`);
+        } else {
+          onEach(data);
+        }
+      });
+    });
+
+    child.on("exit", () => {
+      clearInterval(timer);
+      logger.info(`Finished: ${crawlersDir}/${crawlerFileName} at ${new Date()} in ${(Date.now() - startDate) / 1000}`);
+      callback(null);
+    });
   });
 };
